Added JWT verification

+++ user-service/app/views/user.py
@@ -1,8 +1,9 @@
+import uuid
 from flask import request, jsonify
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
 from sqlalchemy.exc import SQLAlchemyError, IntegrityError
-from flask_jwt_extended import create_access_token, jwt_required
+from flask_jwt_extended import create_access_token, get_jwt_identity, jwt_required
 
 from app.extensions.database import db
 from app.models import User
@@ -18,12 +19,18 @@ class Users(MethodView):
     Endpoint for managing multiple URLs.
     """
 
+    @jwt_required()
     @blp.arguments(UserQueryArgsSchema, location='query')
     @blp.response(200, UserSchema(many=True))
     def get(self, args):
         """
         Get a list of Users based on query parameters.
         """
+        current_user_id = uuid.UUID(get_jwt_identity())
+        current_user = User.query.get_or_404(current_user_id, description="Your User have not been found")
+        if not current_user.is_admin:
+            abort(403, message="You are not authorized to view this user")
+
         return User.query.filter_by(**args)
 
     @blp.arguments(UserSchema)
@@ -50,20 +57,33 @@ class UserById(MethodView):
     """
     Endpoint for managing a single users by its id.
     """
-
-    @blp.response(200, UserSchema)
+    @jwt_required()
+    @blp.response(200, BasicUserSchema)
     def get(self, id):
         """
         Get details of a User by its id.
         """
+        current_user_id = uuid.UUID(get_jwt_identity())
+
+        if current_user_id != id:
+            current_user = User.query.get_or_404(id, description="Your User have not been found")
+            if not current_user.is_admin:
+                abort(403, message="You are not authorized to view this user")
+
         user = User.query.get_or_404(id, description="User not found")
         return user
     
-    @blp.response(200, UserSchema)
+    @jwt_required()
+    @blp.response(200, BasicUserSchema)
     def put(self, id):
         """
         Update details of a User  by its id.
         """
+        current_user_id = uuid.UUID(get_jwt_identity())
+
+        if current_user_id != id:
+            abort(403, message="You are not authorized to modify this user")
+
         payload = request.get_json()
         user = User.query.filter_by(id=id).first_or_404(description="User not found")
 
@@ -78,13 +98,19 @@ class UserById(MethodView):
             message = [str(x) for x in e.args]
             abort(500, message=e.__class__.__name__, errors=message)
 
-        return id
+        return user
     
-    @blp.response(204)
+    @jwt_required()
+    @blp.response(204, BasicUserSchema)
     def delete(self, id):
         """
         Delete a User by its id.
         """
+        current_user_id = uuid.UUID(get_jwt_identity())
+
+        if current_user_id != id:
+            abort(403, message="You are not authorized to delete this user")
+        
         user = User.query.filter_by(id=id).first_or_404(description="User not found")
         try:
             db.session.delete(user)
